Create OTPwithoddsquares.py
Take input from user space separated integers with both even and odd numbers. Generate a 4 digit OTP by using square of odd numbers and trucating the results. If 4 digit OTP is not possible, output -1.
Example: 1 2 5 4 7 Output: 1254 Here odd numbers are 1 5 and 7. There squares are 1 25 and 49. So concatenating and taking the first four digits, we obtain 1254.